move the `ActivateError` to the `errors.rs` and fmt
